+++ src/components/index.tsx
@@ -1,0 +1,7 @@
+import CcInfoBox from "./CcInfoBox";
+import InfoBox from "./InfoBox";
+import Header from "./Header";
+import ScriptContainer from "./ScriptContainer";
+import Loading from "./Loading";
+
+export { CcInfoBox, InfoBox, Header, ScriptContainer, Loading };
