+++ src/pages/index.tsx
@@ -1,0 +1,8 @@
+import MainPage from "./MainPage";
+import ResultPage from "./ResultPage";
+import Login from "./Login";
+import Signup from "./Signup";
+import Mypage from "./Mypage";
+import ErrorPage from "./Errorpage";
+
+export { MainPage, ResultPage, Login, Signup, Mypage, ErrorPage };
